Fix Windows 64 build.
R=verwaest@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/11238060

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12800 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/json-stringifier.h
@@ -587,7 +587,7 @@ void BasicJsonStringifier::SerializeStringUnchecked_(const SrcChar* src,
   }
 
   *(dest++) = '"';
-  current_index_ += dest - dest_start;
+  current_index_ += static_cast<int>(dest - dest_start);
 }
